consolidate transaction and attempt controllers
