+++ web/router.go
@@ -169,10 +169,8 @@ func v2Routes(app services.Application, engine *gin.Engine) {
 		cc := ConfigController{app}
 		authv2.GET("/config", cc.Show)
 
-		tas := TxAttemptsController{app}
-		authv2.GET("/txattempts", tas.Index)
-
 		txs := TransactionsController{app}
+		authv2.GET("/transactions", txs.Index)
 		authv2.GET("/transactions/:TxHash", txs.Show)
 
 		bdc := BulkDeletesController{app}
